requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
PyGithub
faiss-cpu
aiofiles
sentence_transformers
huggingface_hub[hf_xet]
python-multipart
transformers 
torch
hf_xet